back/requirements.txt
django
djangorestframework
django-cors-headers
drf-spectacular
drf-spectacular-sidecar